+++ x/wasm/keeper/genesis_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	wasmvm "github.com/CosmWasm/wasmvm/v2"
+	wasmvm "github.com/CosmWasm/wasmvm/v3"
 	abci "github.com/cometbft/cometbft/abci/types"
 	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	dbm "github.com/cosmos/cosmos-db"
@@ -697,6 +697,7 @@ func setupKeeper(t *testing.T) (*Keeper, sdk.Context) {
 		types.VMConfig{},
 		AvailableCapabilities,
 		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		nil,
 	)
 	return &srcKeeper, ctx
 }