+++ x/wasm/keeper/ibc.go
@@ -20,5 +20,20 @@ func ContractFromPortID(portID string) (sdk.AccAddress, error) {
 	if !strings.HasPrefix(portID, portIDPrefix) {
 		return nil, errorsmod.Wrapf(types.ErrInvalid, "without prefix")
 	}
+
 	return sdk.AccAddressFromBech32(portID[len(portIDPrefix):])
+}
+
+// The port prefix refers to "CosmWasm over IBC v2" and ensures packets are routed to the right entry points
+const portIDPrefixV2 = "wasm2"
+
+func PortIDForContractV2(addr sdk.AccAddress) string {
+	return portIDPrefixV2 + addr.String()
+}
+
+func ContractFromPortID2(portID string) (sdk.AccAddress, error) {
+	if !strings.HasPrefix(portID, portIDPrefixV2) {
+		return nil, errorsmod.Wrapf(types.ErrInvalid, "without prefix")
+	}
+	return sdk.AccAddressFromBech32(portID[len(portIDPrefixV2):])
 }