+++ x/wasm/keeper/testdata/contracts.go
@@ -3,7 +3,7 @@ package testdata
 import (
 	_ "embed"
 
-	typwasmvmtypes "github.com/CosmWasm/wasmvm/v2/types"
+	typwasmvmtypes "github.com/CosmWasm/wasmvm/v3/types"
 
 	"github.com/cosmos/cosmos-sdk/types"
 )